Refactor ISODate and create a new DateUtils class
Extract the static functionality of ISODate to an DateUtils utility class
in common project.
